Refactor parseArguments for the bump automation
