+++ .ci/bumpStackVersion.groovy
@@ -174,16 +174,7 @@ def createPullRequest(Map args = [:]) {
   }
 
   if (bumpUtils.areChangesToBePushed("${args.branchName}")) {
-    def arguments = [
-      title: "${args.title}", labels: "${args.labels}", description: "${args.message}", base: "${args.branchName}"
-    ]
-    if (args.assign?.trim()) {
-      arguments['assign'] = args.assign
-    }
-    if (args.reviewer?.trim()) {
-      arguments['reviewer'] = args.reviewer
-    }
-    githubCreatePullRequest(arguments)
+    githubCreatePullRequest(bumpUtils.parseArguments(args))
   } else {
     log(level: 'INFO', text: "There are no changes to be submitted.")
   }
