+++ vars/bumpUtils.txt
@@ -3,6 +3,7 @@ Utils class for the bump automation pipelines
 * `areChangesToBePushed` -> if there any changes in the existing location to be pushed.
 * `createBranch` -> create a branch given the prefix and suffix arguments. Branch contains the current timestamp.
 * `isVersionAvailable` -> if the given elastic stack version is available.
+* `parseArguments` -> parse the given arguments.
 * `prepareContext` -> prepare the git context, checkout and git config user.name.
 * `getCurrentMinorReleaseFor7` -> retrieve the LATEST known minor release for the 7 major version of the Elastic Stack.
 * `getCurrentMinorReleaseFor6` -> retrieve the LATEST known minor release for the 6 major version of the Elastic Stack.
